SymbolTable: Fix data field, use **Symbol instead of *Symbol
I first tried to limit allocations by directly creating an array of Symbols.
However, this does not work well with the fact that Symbols point to the
next symbol. This was creating infinite loops, so I have to use more
alloallocations in order to allow "next elements". If this poses a
problem, it could be avoided (at the cost of performance) by using a
different hashing algorithm or a different way to order the table